Move projects.test.ts to getProjects.test.ts
