+++ packages/config/src/projects/project/getProjects.ts
@@ -96,6 +96,7 @@ function layer2Or3ToProject(p: Layer2 | Layer3): BaseProject {
     scalingDa: p.dataAvailability,
     tvlInfo: {
       associatedTokens: p.config.associatedTokens ?? [],
+      escrows: p.config.escrows,
       warnings: [p.display.tvlWarning].filter((x) => x !== undefined),
     },
     tvlConfig: getTvlConfig(p),
@@ -186,6 +187,7 @@ function bridgeToProject(p: Bridge): BaseProject {
     bridgeRisks: p.riskView,
     tvlInfo: {
       associatedTokens: p.config.associatedTokens ?? [],
+      escrows: p.config.escrows,
       warnings: [],
     },
     tvlConfig: getTvlConfig(p),
